fix some stuff
