+++ .gitignore
@@ -9,5 +9,4 @@ report
 .idea
 package-lock.json
 /dist/
-/lib/
 .github
